fix: remove assembly name for PROVIDED_AL2 (#77)
When the Lambda runtime is set to `PROVIDE_AL2` the bundling will set
`/p:AssemblyName=bootstrap` to the `dotnet publish` command. However
when a project has project references, this condition will also apply to
the referenced projects resulting in an error `Ambiguous project name
'bootstrap'`.

This fix removes the automatic setting the assembly name. When using
this construct to deploy a custom runtime, make sure that the output
assembly name is `bootstrap` by either add it to the bundling options
`msbuildParameters: ['/p:AssemblyName=bootstrap']` or change the csproj
file to include `<AssemblyName>bootstrap</AssemblyName>`

Co-authored-by: Vincent Lesierse <vinles@amazon.nl>